requirements.txt
numpy==1.23.5
tqdm==4.64.1
rouge_score==0.1.2
pyrouge==0.1.3
jupyterlab==3.5.1
pandas==1.5.2
ipywidgets==8.0.3
scipy==1.9.3
gensim==4.2.0
bert-score==0.3.12
moverscore @ git+https://github.com/AIPHES/emnlp19-moverscore.git@9c362cc5aea61270e988ea0870bf5ae495cc80a3evaluate==0.4.0
pyemd==0.5.1
